Bug and Style Fixes, Broken Registration
Fixed the review step slider so that it starts in a useful range, made
some simplifications to code, and realized that the registration step
was not being implemented in the subtraction algorithm for some reason.
This may have to do with registration producing a transform instead of a
new node. The goal now is to either find a way to A) modify registration
so that it produces a new node instead of a transform (not very useful)
B) apply transforms permanently to nodes or C) make the subtraction
algorithm apply to a transformed node.