Registration Fix, Module Naming Changed
Registration now creates a new volume, instead of applying a transform.
Will likely run into the issue of too many new volumes created to be
useful. TO-DO: Create better volume-rendering options, perhaps add an
Editor/Eraser tool to the 'Review' Step, Make sure that steps can (or
can not) be travelled backward, add a "Reset" button to the review page.