Partway through interactivity changes
I've added several options for customization to the beginning steps.
Further options may be to replace the Thresholding step with an Editor
step, is possible, and either replace the ROI step with a fiducial-based
ROI program OR remove it entirely. Visualization at the end still need
to be functional - would be nice to better integrate the Volume
Rendering module.. Module may be currently non-functional; uploading for
portability.