Major Update, Convex ROI, UI Changes
The biggest changes have to do with moving on from a cubic ROI to a
convex ROI. This uses the Volume Clip with Model module from Slicer.
There are currently additional questions about the normalization process
-- it will likely have to be reconfigured somehow, and it is currently
not functioning. The UI has become messier in this iteration, and will
need cleaning. The ability to go from the first step to the last in a
consistent, clean way is still up in the air, and currently
non-functional.